requirements.txt
aiogram==2.25.1
python-dotenv==1.0.0
SQLAlchemy==2.0.23
aiohttp==3.8.6
Pillow==9.5.0
python-dateutil==2.8.2

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-dependency==0.5.1
pytest-xdist==3.3.1
pytest-cov==4.1.0
colorama==0.4.6

# Development dependencies
black==23.3.0  # Code formatting
flake8==6.0.0  # Linting 